# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nebula-mgmt API
  version: "1.0.0"
  description: >
    Control-plane REST API consumed by the CLI, the nebula-agent poller, and
    mobile bundles. Agent endpoints (enroll, agent/updates) are unauthenticated
    at the bearer layer — enroll carries a single-use token, polls carry an
    Ed25519 proof-of-possession signature (ADR 0004). Admin endpoints require an
    operator API key.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/networks/{id}:
    get:
      operationId: getNetwork
      summary: Get a network by id.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Network.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/mesh-imports:
    get:
      operationId: listMeshImports
      summary: List existing-mesh import sessions visible to the operator.
      responses:
        "200":
          description: Import sessions. Bootstrap token hashes are never returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeshImport"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      operationId: createMeshImport
      summary: Start collecting an existing Nebula mesh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeshImportRequest"
      responses:
        "201":
          description: Session created. The reusable bootstrap token is returned only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportTokenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}:
    get:
      operationId: getMeshImport
      summary: Get an existing-mesh import session without its bootstrap token.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Import session with a deterministic reconciliation preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/rotate-token:
    post:
      operationId: rotateMeshImportToken
      summary: Replace a collecting session's reusable bootstrap token.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "201":
          description: Token rotated. The new token is returned only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImportTokenResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/cancel:
    post:
      operationId: cancelMeshImport
      summary: Cancel a collecting import and remove only its temporary hosts.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Canceled import session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImport"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/mesh-imports/{id}/finalize:
    post:
      operationId: finalizeMeshImport
      summary: Atomically accept the current reconciliation preview and begin managed operation.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeshImportFinalizeRequest"
      responses:
        "200":
          description: Finalized import session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeshImport"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

# Outbound webhooks the server POSTs to an operator-configured endpoint (#256).
# Each delivery carries X-Nebula-Event (type), X-Nebula-Delivery (id), and
# X-Nebula-Signature (sha256=<hmac> over the raw body). The handler/scanner emit
# shapes are validated against these schemas by the contract tests.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Operator API key as a bearer token.
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AgentImportProfile:
      type: object
      additionalProperties: false
      required: [nebula_config_path, nebula_ca_path, nebula_cert_path, nebula_key_path, config_ack_v1]
      properties:
        nebula_config_path: { type: string }
        nebula_ca_path: { type: string }
        nebula_cert_path: { type: string }
        nebula_key_path: { type: string }
        config_ack_v1: { type: boolean }
    AgentImportFirewallRule:
      type: object
      additionalProperties: false
      required: [port, proto, group]
      properties:
        port: { type: string }
        proto: { type: string }
        group: { type: string }
    AgentImportFirewallPolicy:
      type: object
      additionalProperties: false
      required: [inbound, outbound]
      properties:
        inbound:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/AgentImportFirewallRule" }
        outbound:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/AgentImportFirewallRule" }
...
    Network:
      type: object
      required: [id, name, cidrs, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        cidrs:
          type: array
          items:
            type: string
        ca_id:
          type: string
        created_at:
          type: string
          format: date-time
...
    CreateMeshImportRequest:
      type: object
      required: [network_id, ca_id]
      additionalProperties: false
      properties:
        network_id:
          type: string
        ca_id:
          type: string
        expected_hosts:
          type: integer
          minimum: 1
    MeshImport:
      type: object
      required:
        - id
        - network_id
        - ca_id
        - owner_operator_id
        - ca_fingerprint
        - status
        - revision
        - token_expires_at
        - captured_network_config_version
        - created_at
        - updated_at
      properties:
        id:
          type: string
        network_id:
          type: string
        ca_id:
          type: string
        owner_operator_id:
          type: string
        ca_fingerprint:
          type: string
        status:
          type: string
          enum: [collecting, finalized, canceled]
        expected_hosts:
          type: integer
          minimum: 1
        revision:
          type: integer
          format: int64
          minimum: 0
        token_expires_at:
          type: string
          format: date-time
        captured_network_config_version:
          type: integer
          minimum: 0
        terminal_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        finalized_at:
          type: string
          format: date-time
        canceled_at:
          type: string
          format: date-time
    MeshImportIssue:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        snapshot_id: { type: string }
        host_name: { type: string }
        field: { type: string }
        message: { type: string }
    MeshImportHostProposal:
      type: object
      required: [snapshot_id, profile, host]
      properties:
        snapshot_id: { type: string }
        profile: { $ref: "#/components/schemas/AgentImportProfile" }
        host: { $ref: "#/components/schemas/Host" }
    MeshImportProposal:
      type: object
      required: [hosts, firewall, blocklist]
      properties:
        hosts:
          type: array
          items: { $ref: "#/components/schemas/MeshImportHostProposal" }
        firewall: { $ref: "#/components/schemas/AgentImportFirewallPolicy" }
        blocklist:
          type: array
          items: { type: string }
    MeshImportReport:
      type: object
      required: [blockers, warnings, proposal]
      properties:
        blockers:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/MeshImportIssue" }
        warnings:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/MeshImportIssue" }
        proposal: { $ref: "#/components/schemas/MeshImportProposal" }
    MeshImportWarningAcknowledgement:
      type: object
      required: [key, issue]
      properties:
        key:
          type: string
          pattern: ^[a-f0-9]{64}$
        issue: { $ref: "#/components/schemas/MeshImportIssue" }
    MeshImportDetail:
      type: object
      required: [mesh_import, host_count, expected_count_satisfied, report, warning_acknowledgements]
      properties:
        mesh_import: { $ref: "#/components/schemas/MeshImport" }
        host_count: { type: integer, minimum: 0 }
        expected_count_satisfied: { type: boolean }
        report: { $ref: "#/components/schemas/MeshImportReport" }
        warning_acknowledgements:
          type: array
          items: { $ref: "#/components/schemas/MeshImportWarningAcknowledgement" }
    MeshImportFinalizeRequest:
      type: object
      additionalProperties: false
      required: [revision, inventory_complete]
      properties:
        revision: { type: integer, format: int64, minimum: 0 }
        inventory_complete: { type: boolean }
        acknowledged_warnings:
          type: [array, "null"]
          items:
            type: string
            pattern: ^[a-f0-9]{64}$
    MeshImportTokenResponse:
      type: object
      required: [mesh_import, token]
      properties:
        mesh_import:
          $ref: "#/components/schemas/MeshImport"
        token:
          type: string
          pattern: ^nmi_[A-Za-z0-9_-]+$
    Host:
      type: object
      required:
        - id
        - network_id
        - name
        - nebula_ips
        - groups
        - role
        - is_lighthouse
        - is_relay
        - status
        - kind
        - created_at
        - updated_at
      properties:
        id:
          type: string
        network_id:
          type: string
        ca_id:
          type: string
        name:
          type: string
        nebula_ips:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        role:
          type: string
        is_lighthouse:
          type: boolean
        is_relay:
          type: boolean
        public_ip:
          type: string
        listen_port:
          type: integer
        status:
          type: string
        cert_fingerprint:
          type: string
        prev_cert_fingerprint:
          type: string
        cert_expires_at:
          type: string
          format: date-time
        cert_rotated_at:
          type: string
          format: date-time
        pending_rekey:
          type: boolean
        signing_pub_pem:
          type: string
        last_seen_at:
          type: string
          format: date-time
        kind:
          type: string
        variant:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...